adding modularity for the project , split each core class into a core classes folder , and split the commands into a dedicated folder
